fix some bugs

+++ socket.h
@@ -1,8 +1,8 @@
 #ifndef _SOCKET_H_
 #define _SOCKET_H_
 
-#include <stdint.h>
-#include "url.h"
+#include <sys/types.h>
+#include <sys/socket.h>
 
 #ifndef INET_ADDRSTRLEN
 #define INET_ADDRSTRLEN		16	/* xxx.xxx.xxx.xxx\0 */
@@ -12,12 +12,12 @@
 #define UDP	SOCK_DGRAM
 
 int socket_create(int type);
-void socket_bind(int sockfd, unsigned short port);
+void socket_bind(int sockfd, int port);
 void socket_set_non_blocking(int sockfd);
 void socket_start_listening(int sockfd);
-void tcp_server_init(uint16_t port);
-int socket_connect(const char *host, uint16_t port);
+void tcp_server_init(int port);
+int socket_connect(const char *host, int port);
 int socket_recv(int sockfd, void *buff, int size);
 int socket_send(int sockfd, const void *buff, int size);
 
-#endif	/* _SOCKET_H_ */
+#endif /* _SOCKET_H_ */
